Added spdlog sorted out formatting & includes made example nodes actually do things

+++ src/common/skylight_time.h
@@ -1,9 +1,10 @@
 #pragma once
+
 #include <chrono>
 
 namespace skylight {
-    uint64_t Now() {
+    int64_t Now() {
         std::chrono::duration d = std::chrono::high_resolution_clock::now().time_since_epoch();
-        return std::chrono::duration_cast<std::chrono::microseconds>(d).count();
+        return static_cast<int64_t>(std::chrono::duration_cast<std::chrono::microseconds>(d).count());
     }
 }
